refactor: move scatterView in own dir
