+++ frontend/src/instance-views/scatterView/NoEmbed.svelte
