+++ frontend/src/instance-views/scatterView/reglScatter/ReglScatter.svelte
